+++ beatmap/v3/slider.ts
@@ -295,7 +295,8 @@ export class Slider extends BaseSlider<ISlider> {
             this.posX < 0 ||
             this.posX > 3 ||
             this.tailPosX < 0 ||
-            this.tailPosX > 3
+            this.tailPosX > 3 ||
+            (this.posX === this.tailPosX && this.posY === this.tailPosY && this.time === this.tailTime)
         );
     }
 }